Added additional condition in functions 'follow_user' and 'unfollow_user' in case if user with gotten 'pk' doesn't exist.

+++ django_net_core/applications/frontend/views.py
@@ -19,6 +19,7 @@ from applications.user_wall.services.utils import form_utils as uw_form_utils
 
 from applications.groups import forms as g_forms
 from applications.groups.permissions import GROUP_FORBIDDEN_MESSAGE, GROUP_CREATION_FORBIDDEN_MESSAGE
+from applications.groups.services import utils as g_utils
 from applications.groups.services.crud import create as g_create, read as g_read
 
 
@@ -110,6 +111,9 @@ def follow_user(request: WSGIRequest, pk: int):
         return redirect(to='login')
 
     owner = up_read.get_raw_user_instance(user_pk=pk)
+    if not owner:
+        raise Http404
+
     if not up_common_utils.is_followed(current_user=owner, visitor=request.user):
         up_create.create_new_follower(owner=owner, follower=request.user)
     return redirect(to='user_profile', pk=pk)
@@ -120,6 +124,9 @@ def unfollow_user(request: WSGIRequest, pk: int):
         return redirect(to='login')
 
     owner = up_read.get_raw_user_instance(user_pk=pk)
+    if not owner:
+        raise Http404
+
     if up_common_utils.is_followed(current_user=owner, visitor=request.user):
         up_delete.delete_follower(owner=owner, follower=request.user)
     return redirect(to='user_profile', pk=pk)
@@ -234,9 +241,20 @@ class GroupPostCreationView(View):
         pass
 
 
-def create_group_post(request: WSGIRequest) -> HttpResponse:
-    if not request.method.lower() == 'post':
-        return HttpResponseForbidden(GROUP_CREATION_FORBIDDEN_MESSAGE)
+def follow_group(request: WSGIRequest, group_slug: str):
+    if not request.user.is_authenticated:
+        return redirect(to='login')
+
+    group = g_read.get_group_by_slug(group_slug)
+
+
+def unfollow_group(request: WSGIRequest, group_slug: str):
+    pass
+
+
+# def create_group_post(request: WSGIRequest) -> HttpResponse:
+#     if not request.method.lower() == 'post':
+#         return HttpResponseForbidden(GROUP_CREATION_FORBIDDEN_MESSAGE)
 
 
 class UserWallView(LoginRequiredMixin, View):
